Add ability to type form component (#4841)
